requirements.txt
fastapi~=0.115.0
uvicorn[standard]~=0.35.0
sqlalchemy~=2.0.41
pydantic[email]~=2.11.7
python-jose[cryptography]~=3.3.0
passlib[bcrypt]~=1.7.4
python-multipart~=0.0.9
email-validator~=2.1.1
pycryptodome~=3.20.0
requests~=2.32.3
python-dotenv~=1.0.1
httpx~=0.27.2
bleach~=6.2.0
beautifulsoup4~=4.12.3
soupsieve~=2.6
dnspython~=2.7.0
idna~=3.10
httpcore~=1.0.6
cryptography~=45.0.4
anyio~=4.6.2.post1
h11~=0.14.0
defusedxml~=0.7.1
ipython~=8.29.0
numpy~=2.1.0
pillow~=11.0.0
jupyter~=1.1.1
pip~=25.1.1
attrs~=24.2.0
rsa~=4.9.1
pyasn1~=0.6.1
pyzmq~=26.2.0
colorama~=0.4.6
tornado~=6.4.1
cffi~=1.17.1
setuptools~=75.2.0
comm~=0.2.2
traitlets~=5.14.3
fqdn~=1.5.1
parso~=0.8.4
jedi~=0.19.1
ecdsa~=0.19.1
pytz~=2024.1
PyYAML~=6.0.2
sniffio~=1.3.1
arrow~=1.3.0
python-dateutil~=2.9.0.post0
babel~=2.16.0
typing_extensions~=4.14.0
six~=1.16.0
click~=8.2.1
Pygments~=2.18.0
certifi~=2024.8.30
pywin32~=306
charset-normalizer~=3.4.0
overrides~=7.7.0
autobahn~=24.4.2
txaio~=23.1.1
argon2-cffi-bindings~=21.2.0
webencodings~=0.5.1
tinycss2~=1.4.0
MarkupSafe~=3.0.2
psutil~=6.1.0
pywinpty~=2.0.14
debugpy~=1.8.7
greenlet~=3.2.3
decorator~=4.4.2
starlette~=0.46.2
email_validator~=2.1.1
pydantic_core~=2.33.2
websockets~=15.0.1
stack-data~=0.6.3
packaging~=24.1
matplotlib~=3.10.1
matplotlib-inline~=0.1.7
jupyter_client~=8.6.3
prompt_toolkit~=3.0.48
wcwidth~=0.2.13
urllib3~=2.2.3
httptools~=0.6.4
watchfiles~=1.1.0
Jinja2~=3.1.4
PyNaCl~=1.5.0
hyperlink~=21.0.0
nbformat~=5.10.4
jupyter_core~=5.7.2
nbclient~=0.10.0
fastjsonschema~=2.20.0
jsonschema~=4.23.0
jupyter_server~=2.14.2
jupyterlab~=4.2.5
jupyterlab_server~=2.27.3
notebook_shim~=0.2.4
notebook~=7.2.2
annotated-types~=0.7.0
asttokens~=2.4.1
contourpy~=1.3.2
executing~=2.1.0
fonttools~=4.57.0
scipy~=1.15.3
ipykernel~=6.29.5
nbconvert~=7.16.4
pandocfilters~=1.5.1
mistune~=3.0.2
jupyterlab_pygments~=0.3.0
pure_eval~=0.2.3
pycparser~=2.22
pyparsing~=3.2.3
websocket-client~=1.8.0
terminado~=0.18.1
ipywidgets~=8.1.5
referencing~=0.35.1
webcolors~=24.8.0
jsonpointer~=3.0.0
uri-template~=1.3.0
isoduration~=20.11.0
rfc3339-validator~=0.1.4
jsonschema-specifications~=2024.10.1
async-lru~=2.0.4
cycler~=0.12.1
kiwisolver~=1.4.8
Send2Trash~=1.8.3
platformdirs~=4.3.6
jupyter-lsp~=2.2.5
jupyter-events~=0.10.0
python-json-logger~=2.0.7
prometheus_client~=0.21.0
jupyter_server_terminals~=0.5.3
jupyter-console~=6.6.3
json5~=0.9.25